+++ client/src/components/Query/Statement.jsx
@@ -19,7 +19,7 @@ import {
   INSTRUCTION_TYPE_SUBQUERY, SUBQUERY_TYPE_INTERSECT, SUBQUERY_TYPE_UNITE, SUBQUERY_TYPE_SUBTRACT, createSubquery,
 } from './Subquery';
 import { createOperator } from './Operator';
-import VariantTable from './VariantTable'
+
 
 const MAX_QUERIES = 15;
 const MAX_REVISIONS = 10;
@@ -649,8 +649,6 @@ class Statement extends React.Component {
             />
           ) : queries
         }
-        <br/>
-        <VariantTable/>
       </div>
     );
   }
