+++ client/src/components/screens/PatientVariant/index.jsx
@@ -13,7 +13,8 @@ import Header from '../../Header';
 import Navigation from '../../Navigation';
 import Content from '../../Content';
 import Footer from '../../Footer';
-import VariantNavigation from './components';
+import VariantNavigation from './components/VariantNavigation';
+import VariantResultsTable from './components/VariantResultsTable';
 
 import './style.scss';
 import { patientShape } from '../../../reducers/patient';
@@ -129,6 +130,12 @@ class PatientVariantScreen extends React.Component {
               onEditCallback={this.handleQueryChange}
               onRemoveCallback={this.handleQueryRemoval}
               onDuplicateCallback={this.handleQueryDuplication}
+            />
+            <br/>
+            <br />
+            <VariantResultsTable
+                key="variant-results"
+                intl={intl}
             />
         </Card>
         <Footer />
